+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fix suggestion module name
 - fix '\t' in linebreak
 - fix `PATH` not in root env
+- pprint differ heads
 
 ## [0.7.1]
 - rename `List.some` and `List.every` to `any` and `all`